pyproject.toml
[tool.poetry]
name = "carberretta"
version = "0.1.1"
description = "Discord bot designed to handle specific tasks in the Carberra Tutorials server."
authors = ["Ethan Henderson <parafoxia@carberra.xyz>", "Max Shaughnessy <max@mshaugh.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.3"
apscheduler = "^3.6.3"
aiosqlite = "^0.13.0"
aiofiles = "^0.5.0"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
